requirements.txt
aniso8601==10.0.0
black==25.1.0
blinker==1.9.0
click==8.1.8
flake8==7.1.2
Flask==3.1.0
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pytz==2025.1
six==1.17.0
Werkzeug==3.1.3
